+++ CHANGELOG.md
@@ -35,3 +35,11 @@ Added book chapters content fetching for faster performance
 ### Features
 - Search books
 - Get book by url
+
+## [0.0.19]
+convert book to epub(beta)
+
+### Features
+- Search books
+- Get book by url
+- Convert books to epub(beta)